recombined figure files. base plots multiple figs if present
