+++ mrsa_ca_rna/data/Tier1.csv
@@ -1,0 +1,21 @@
+﻿GLS2
+HLA-DRA
+CD74
+DNMT3A
+IL10
+STAT1
+CEBPB
+TIRAP
+IL1RN
+IL27
+CHI3L1
+CCL2
+ANGPT2
+CCL26
+CXCL8
+sele
+IL2RA
+LCN2
+IL1B
+IL6
+CXCL1
