+++ mrsa_ca_rna/data/Tier2.csv
@@ -1,0 +1,93 @@
+﻿HLA-DRB1
+TLR1
+TLR2
+TLR6
+TLR10
+TNF
+IL17A
+IFNG
+MPO
+G6PD
+ITGB2
+SLC35C1
+FERMT3
+STAT3
+IRAK4
+MyD88
+IKBKG
+NFKBIA
+RAC2
+PGM3
+HLA-DMA
+HLA-DMB
+HLA-DOA
+HLA-DOB
+HLA-DPA1
+HLA-DPA2
+HLA-DPA3
+HLA-DPB1
+HLA-DPB2
+HLA-DQA1
+HLA-DQA1-AS1
+HLA-DQA2
+HLA-DQB1
+HLA-DQB1-AS1
+HLA-DQB2
+HLA-DQB3
+HLA-DRB2
+HLA-DRB3
+HLA-DRB4
+HLA-DRB5
+HLA-DRB6
+HLA-DRB7
+HLA-DRB8
+HLA-DRB9
+NOX1
+NOX3
+NOX4
+NOX4P1
+NOX5
+NOXA1
+NOXO1
+ANGPT1
+CCL5
+IL15
+IL18
+IL6R
+TNFRSF1A
+TNFRSF1B
+TEK
+TNFSF10
+TREM1
+C5
+CXCL10
+CSF2
+IL1A
+LIF
+CSF1
+PLAUR
+IL4
+IL12A
+IL12B
+COL18A1
+GLS2
+HLA-DRA
+CD74
+DNMT3A
+IL10
+STAT1
+CEBPB
+TIRAP
+IL1RN
+IL27
+CHI3L1
+CCL2
+ANGPT2
+CCL26
+CXCL8
+sele
+IL2RA
+LCN2
+IL1B
+IL6
+CXCL1
