+++ mrsa_ca_rna/data/Tier3.csv
@@ -1,0 +1,815 @@
+﻿ADAM8
+AMH
+FASLG
+BDNF
+BID
+C5
+CASP3
+CASP8
+TNFSF8
+CD40LG
+CD70
+CDH5
+CISH
+CNTF
+CSF1
+CSF2
+CSF3
+CSH1
+CSH2
+CSHL1
+CTF1
+CX3CR1
+DDT
+DEFB1
+DEFB4A
+ECM1
+EDA
+EFNA5
+ENG
+EPO
+VEGFD
+FKBP1A
+GATA3
+GH1
+GH2
+GRB2
+CXCL1
+CXCL2
+CXCL3
+IFNA1
+IFNA2
+IFNA4
+IFNA5
+IFNA6
+IFNA7
+IFNA8
+IFNA10
+IFNA13
+IFNA14
+IFNA16
+IFNA17
+IFNA21
+IFNB1
+IFNG
+IFNW1
+IL1A
+IL1B
+IL1RN
+IL2
+IL3
+IL4
+IL5
+IL6
+IL6R
+IL6ST
+IL7
+CXCL8
+IL9
+IL10
+IL11
+IL12A
+IL12B
+IL12RB1
+IL13
+IL15
+IL18
+INHBC
+CXCL10
+ITGA5
+ITGB3
+JAK1
+JAK2
+JAK3
+LIF
+LIFR
+LTA
+LTB
+SMAD2
+SMAD3
+SMAD6
+SMAD7
+KITLG
+MIF
+CXCL9
+TRIM37
+MYD88
+NARS1
+NGF
+NTF3
+NTF4
+NTRK1
+OSM
+PF4
+PF4V1
+PGF
+PIK3R1
+PLCG1
+PPBP
+PRL
+CCL1
+CCL2
+CCL3
+CCL3L1
+CCL4
+CCL5
+CCL7
+CCL8
+CCL11
+CCL13
+CCL14
+CCL15
+CCL16
+CCL17
+CCL18
+CCL19
+CCL20
+CCL21
+CCL22
+CCL23
+CCL24
+CCL25
+CXCL6
+CXCL11
+CXCL5
+XCL1
+CX3CL1
+SDCBP
+CXCL12
+SHC1
+STAT1
+XCL2
+SYK
+ADAM17
+MAP3K7
+TFF2
+TGFB1
+TGFB2
+TGFB3
+LEFTY2
+TGFBR1
+TGFBR2
+TGFBR3
+TLR5
+TNF
+TRAF1
+TRAF2
+TRAF3
+TRAF6
+TRIP6
+TNFSF4
+TYK2
+VEGFA
+VEGFB
+VEGFC
+PXDN
+YARS1
+MADD
+TNFSF11
+SOCS1
+TRADD
+RIPK1
+TNFSF14
+TNFSF13
+TNFSF12
+TNFSF10
+TNFSF9
+FADD
+SOCS2
+CFLAR
+TNFSF18
+NOL3
+SOCS3
+CCRL2
+DNAJA3
+OSMR
+CRLF1
+TGFBRAP1
+CXCL14
+BABAM2
+TRAF4
+USP15
+TNFSF15
+CASP8AP2
+SH2B3
+FEM1B
+TRAP1
+EBI3
+RNF41
+CCL26
+PIBF1
+CREB3
+CXCL13
+SIVA1
+LEFTY1
+TNFSF13B
+NES
+FRS2
+CCL27
+FERMT2
+CLCF1
+BAMBI
+STAP1
+IL36RN
+GREM1
+BEX3
+IL36B
+IL37
+IL36A
+CNIH4
+PYCARD
+IL20
+IL22
+IRAK4
+CKLF
+IL23A
+ERAP1
+TLR9
+TOLLIP
+PIDD1
+CCDC88A
+DEFB103B
+IL36G
+CCL28
+IFNK
+SMURF1
+S100A14
+CXCL16
+SNX6
+IL21
+TMBIM1
+IL25
+RASL11B
+PDCL3
+ITCH
+IL1F10
+TSLP
+IL33
+TIMM50
+IL17F
+LRG1
+IL34
+CD300LF
+IL23R
+DAB2IP
+SPRED1
+CADM4
+SPRED2
+DEFB106A
+DEFB110
+DEFB114
+DEFB130A
+IL27
+IFNL1
+IFNE
+IL31
+GPR15LG
+CCL4L1
+DEFB133
+CCL3L3
+DEFB103A
+RHEX
+DEFB106B
+DEFB109B
+MSMP
+CCR2
+DEFB130B
+TGFBR3L
+AREG
+BMP1
+BMP2
+BMP3
+BMP4
+BMP5
+BMP6
+BMP7
+BMP8B
+EDN1
+FGF2
+FLT3LG
+GDF1
+GDF2
+MSTN
+GDF9
+GDF10
+GPI
+GRN
+NRG1
+HMGB1
+IL16
+IL17A
+INHA
+INHBA
+INHBB
+NDP
+NODAL
+TNFRSF11B
+SECTM1
+SPP1
+THPO
+TIMP1
+WNT1
+WNT2
+WNT3
+WNT5A
+WNT6
+WNT7A
+WNT7B
+WNT8A
+WNT8B
+WNT10B
+WNT11
+WNT2B
+WNT9A
+WNT9B
+SCG2
+GDF5
+BMP15
+IL32
+AIMP1
+CER1
+ADIPOQ
+GDF15
+GDF3
+NAMPT
+GDF11
+FAM3C
+IL24
+TAFA5
+IL17C
+IL17B
+BMP10
+IL19
+WNT16
+IL17D
+FAM3B
+WNT4
+THNSL2
+IL26
+SLURP1
+GREM2
+WNT10A
+WNT5B
+INHBE
+WNT3A
+SCGB3A1
+CMTM7
+CMTM1
+C1QTNF4
+CMTM5
+CMTM3
+FAM3D
+CMTM2
+GDF7
+CMTM8
+IFNL2
+IFNL3
+VSTM1
+ALKAL2
+BMP8A
+ALKAL1
+GDF6
+C17orf99
+IFNL4
+ANXA11
+ATP1B1
+B2M
+CD4
+MS4A1
+CD74
+CD81
+HLA-DMA
+HLA-DMB
+HLA-DOA
+HLA-DOB
+HLA-DPA1
+HLA-DPB1
+HLA-DQA1
+HLA-DQA2
+HLA-DQB1
+HLA-DQB2
+HLA-DRA
+HLA-DRB1
+HLA-DRB3
+HLA-DRB4
+HLA-DRB5
+HSPA8
+HSP90AA1
+HSP90AB1
+PKM
+YWHAE
+CD8A
+KLRC1
+KLRC2
+KLRD1
+TAPBP
+LILRB2
+LILRB1
+CD160
+CYRIB
+TAPBPL
+DHCR24
+HLA-A
+HLA-B
+HLA-C
+HLA-E
+HLA-F
+HLA-G
+HLA-H
+MICB
+TAP1
+TAP2
+TRGV3
+TRGV9
+SLC7A5
+MAML1
+CLEC4M
+SLC7A9
+SLC7A8
+TRBV28
+TRBV12-3
+TRBV7-9
+TRAV29DV5
+TRAV23DV6
+TRAV19
+TRAV12-3
+TRAV12-2
+TRAV12-1
+TRAV8-4
+CD209
+MICA
+CXCR5
+C3AR1
+C5AR1
+CD44
+CCR1
+CCR3
+CCR4
+CCR5
+CCR6
+CCR7
+CCR8
+ACKR2
+CMKLR1
+CNTFR
+CR1
+CR2
+CSF2RA
+CSF2RB
+CSF3R
+CTSH
+EPOR
+F3
+FCAR
+FCER1A
+FCER1G
+FCER2
+FCGR1A
+FCGR1BP
+FCGR2A
+FCGR2B
+FCGR3A
+FCGR3B
+FLT3
+FPR1
+FPR2
+FPR3
+GFRA1
+GFRA2
+GFRA3
+GHR
+CCR10
+XCR1
+CXCR3
+GPR17
+GPR32
+GPR32P1
+GPR33
+GPR35
+MR1
+IFNAR1
+IFNAR2
+IFNGR1
+IFNGR2
+IL1R1
+IL1RAP
+IL2RA
+IL2RB
+IL2RG
+IL3RA
+IL4R
+IL5RA
+IL7R
+CXCR1
+CXCR2
+IL9R
+IL10RA
+IL10RB
+IL11RA
+IL12RB2
+IL13RA1
+IL13RA2
+IL15RA
+KIR2DL4
+KIR2DS5
+KIR3DL1
+KIR3DS1
+LEPR
+MPL
+PIGR
+PRLR
+IL1R2
+CXCR4
+FZD4
+IL18RAP
+IL1RL2
+IL18R1
+FCGR2C
+IL1RL1
+FCMR
+IL27RA
+CXCR6
+CCR9
+GPR75
+LILRB4
+LILRA1
+LILRA3
+LILRA2
+KLRK1
+LILRA4
+IL17RA
+IL1RAPL2
+C5AR2
+IL21R
+KLRF1
+ACKR4
+IL20RA
+IL20RB
+IL17RD
+IL17RB
+ACKR3
+IL22RA1
+GFRA4
+CRLF2
+IL17RC
+IL22RA2
+CD200R1
+IL17RE
+IL31RA
+IFNLR1
+NCR3
+LILRA5
+GFRAL
+IL17REL
+KLRC4-KLRK1
+ABO
+CD1A
+CD1B
+CD1C
+CD1D
+CD1E
+CD40
+FCN1
+FCN2
+GP2
+IGHA1
+IGHA2
+IGHD
+IGHE
+IGHG1
+IGHG2
+IGHG3
+IGHG4
+IGHM
+JCHAIN
+IGKC
+IGLC1
+IGLC2
+IGLC3
+IGLC6
+IGLL1
+ITGA4
+KIR2DL3
+LAG3
+PLG
+PPP2R1A
+SLAMF1
+TSPAN4
+FCN3
+TOPORS
+SPON2
+IGKV1-5
+IGHV8-51-1
+IGHV7-81
+IGHV6-1
+IGHV5-10-1
+IGHV5-51
+IGHV4-38-2
+IGHV4-61
+IGHV4-59
+IGHV4-39
+IGHV4-34
+IGHV4-31
+IGHV4-30-4
+IGHV4-28
+IGHV4-4
+IGHV3-38-3
+IGHV3-74
+IGHV3-73
+IGHV3-72
+IGHV3-66
+IGHV3-64
+IGHV3-53
+IGHV3-49
+IGHV3-48
+IGHV3-43
+IGHV3-38
+IGHV3-35
+IGHV3-33
+IGHV3-30
+IGHV3-23
+IGHV3-21
+IGHV3-20
+IGHV3-16
+IGHV3-15
+IGHV3-13
+IGHV3-11
+IGHV3-9
+IGHV3-7
+IGHV2-70
+IGHV2-26
+IGHV2-5
+IGHV1-69-2
+IGHV1-38-4
+IGHV1-69
+IGHV1-58
+IGHV1-45
+IGHV1-24
+IGHV1-18
+IGHV1-8
+IGHV1-3
+IGLV7-43
+IGLV6-57
+IGLV3-27
+IGLV3-25
+IGLV3-21
+IGLV3-19
+IGLV3-1
+IGLV2-23
+IGLV2-14
+IGLV2-11
+IGLV2-8
+IGLV1-51
+IGLV1-47
+IGLV1-44
+IGLV1-40
+IGLC7
+IGKV2D-28
+IGKV1D-39
+IGKV1D-33
+IGKV1D-12
+IGKV5-2
+IGKV4-1
+IGKV3-20
+IGKV3-15
+IGKV2-30
+IGKV1-39
+IGKV1-17
+IGKV1-16
+IGHV7-4-1
+IGHV3-30-3
+EP400
+IGHV3-30-5
+SIRPA
+NCR3LG1
+IGLL5
+IGHV2-70D
+IGHV1-69D
+IGHV3-64D
+IGHV3-43D
+CD36
+S100A8
+S100A9
+TLR1
+TLR2
+TLR6
+LY96
+UNC93B1
+TLR10
+TIRAP
+CYBA
+CYBB
+NCF2
+NOX1
+DUOX2
+NOX4
+NOX3
+DUOX1
+NOX5
+NCF1
+FMO5
+KMO
+AIFM1
+MICAL2
+CYB5R4
+MICAL1
+BST1
+CD38
+TLR4
+IL1RAPL1
+SARM1
+A2M
+ACVR1
+ACVRL1
+BGN
+BMPR1A
+BMPR1B
+BMPR2
+CASP1
+COMP
+CSF1R
+ELANE
+ACKR1
+LRRC32
+GBP1
+ITGAV
+ITGB1
+KIT
+LTBP1
+LTBP3
+MMP8
+NBL1
+PLP2
+THBS1
+TNFRSF1A
+TNFRSF1B
+ZFP36
+LTBP4
+TCAP
+CHRD
+HYAL2
+TNFRSF14
+TNFRSF11A
+NRP2
+NRP1
+NOG
+IL18BP
+HAX1
+TRIM16
+PDPN
+COPS5
+PARK7
+CHRDL2
+SOSTDC1
+TSKU
+KLHL20
+SOST
+TWSG1
+CHRDL1
+VASN
+WFIKKN1
+WFIKKN2
+TMC8
+HJV
+NLRP7
+UCMA
+SCUBE3
+NRROS
+MICOS10-NBL1
+MPO
+G6PD
+ITGB2
+SLC35C1
+FERMT3
+STAT3
+IKBKG
+NFKBIA
+RAC2
+PGM3
+HLA-DPA2
+HLA-DPA3
+HLA-DPB2
+HLA-DQA1-AS1
+HLA-DQB1-AS1
+HLA-DQB3
+HLA-DRB2
+HLA-DRB6
+HLA-DRB7
+HLA-DRB8
+HLA-DRB9
+NOX4P1
+NOXA1
+NOXO1
+ANGPT1
+TNFR1
+TEK
+TREM1
+PLAUR
+COL18A1
+GLS2
+DNMT3A
+CEBPB
+CHI3L1
+ANGPT2
+SELE
+LCN2
