+++ mrsa_ca_rna/figures/base.py
@@ -107,7 +107,17 @@ def genFigure():
     nameOut = "figure" + sys.argv[1]
 
     exec("from mrsa_ca_rna.figures import " + nameOut)
-    ff = eval(nameOut + ".genFig()")
-    ff.savefig(fdir + nameOut + ".svg", bbox_inches="tight", pad_inches=0.1)
+    figures = eval(nameOut + ".genFig()")
+
+    # Check if multiple figures were returned
+    if isinstance(figures, list | tuple):
+        for i, fig in enumerate(figures):
+            fig_name = f"{nameOut}_{i+1}"
+            fig.savefig(f"{fdir}{fig_name}.svg", bbox_inches="tight", pad_inches=0.1)
+            logging.info(f"Saved figure {i+1} as {fig_name}.svg")
+    else:
+        # Single figure returned
+        figures.savefig(f"{fdir}{nameOut}.svg", bbox_inches="tight", pad_inches=0.1)
+        logging.info(f"Saved figure as {nameOut}.svg")
 
     logging.info("%s is done after %s seconds.", nameOut, time.time() - start)
